requirements.txt
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.2.0
dnspython==2.7.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.1
MarkupSafe==3.0.2
numpy==2.2.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pymongo==4.11.1
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.29.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
